+++ trunk/platform/a610/main.c
@@ -11,7 +11,7 @@ long get_vbatt_max()
 }
 
 static const int fl_tbl[] = {7300, 8460, 9565, 10835, 12565, 14926, 17342, 21709, 29200};
-#define NUM_FL (sizeof(fl_tbl)/sizeof(fl_tbl[0]))
+#define NUM_FL (int)(sizeof(fl_tbl)/sizeof(fl_tbl[0]))
 #define CF_EFL 47946
 
 const int zoom_points = NUM_FL;
