+++ trunk/platform/s5is/main.c
@@ -56,7 +56,7 @@ static const struct {
 	{ 105,  61400 },
 	{ 128,  72000 },
 };
-#define NUM_FL (sizeof(fl_tbl)/sizeof(fl_tbl[0]))
+#define NUM_FL (int)(sizeof(fl_tbl)/sizeof(fl_tbl[0]))
 #define CF_EFL 6000
 
 const int zoom_points = 129;
